Adressed some comments but still moving the API from common to nested
